backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==3.2.3
django-colorfield==0.11.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
html5lib==1.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
oauthlib==3.2.2
oscrypto==1.3.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
